+++ oci/dataflow_applications_data_source.go
@@ -1,0 +1,165 @@
+// Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
+
+package oci
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform/helper/schema"
+	oci_dataflow "github.com/oracle/oci-go-sdk/dataflow"
+)
+
+func DataflowApplicationsDataSource() *schema.Resource {
+	return &schema.Resource{
+		Read: readDataflowApplications,
+		Schema: map[string]*schema.Schema{
+			"filter": dataSourceFiltersSchema(),
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"display_name_starts_with": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"owner_principal_id": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"applications": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem:     GetDataSourceItemSchema(DataflowApplicationResource()),
+			},
+		},
+	}
+}
+
+func readDataflowApplications(d *schema.ResourceData, m interface{}) error {
+	sync := &DataflowApplicationsDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).dataFlowClient
+
+	return ReadResource(sync)
+}
+
+type DataflowApplicationsDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_dataflow.DataFlowClient
+	Res    *oci_dataflow.ListApplicationsResponse
+}
+
+func (s *DataflowApplicationsDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *DataflowApplicationsDataSourceCrud) Get() error {
+	request := oci_dataflow.ListApplicationsRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if displayNameStartsWith, ok := s.D.GetOkExists("display_name_starts_with"); ok {
+		tmp := displayNameStartsWith.(string)
+		request.DisplayNameStartsWith = &tmp
+	}
+
+	if ownerPrincipalId, ok := s.D.GetOkExists("owner_principal_id"); ok {
+		tmp := ownerPrincipalId.(string)
+		request.OwnerPrincipalId = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(false, "dataflow")
+
+	response, err := s.Client.ListApplications(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	request.Page = s.Res.OpcNextPage
+
+	for request.Page != nil {
+		listResponse, err := s.Client.ListApplications(context.Background(), request)
+		if err != nil {
+			return err
+		}
+
+		s.Res.Items = append(s.Res.Items, listResponse.Items...)
+		request.Page = listResponse.OpcNextPage
+	}
+
+	return nil
+}
+
+func (s *DataflowApplicationsDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(GenerateDataSourceID())
+	resources := []map[string]interface{}{}
+
+	for _, r := range s.Res.Items {
+		application := map[string]interface{}{
+			"compartment_id": *r.CompartmentId,
+		}
+
+		if r.DefinedTags != nil {
+			application["defined_tags"] = definedTagsToMap(r.DefinedTags)
+		}
+
+		if r.DisplayName != nil {
+			application["display_name"] = *r.DisplayName
+		}
+
+		application["freeform_tags"] = r.FreeformTags
+
+		if r.Id != nil {
+			application["id"] = *r.Id
+		}
+
+		application["language"] = r.Language
+
+		if r.OwnerPrincipalId != nil {
+			application["owner_principal_id"] = *r.OwnerPrincipalId
+		}
+
+		if r.OwnerUserName != nil {
+			application["owner_user_name"] = *r.OwnerUserName
+		}
+
+		application["state"] = r.LifecycleState
+
+		if r.TimeCreated != nil {
+			application["time_created"] = r.TimeCreated.String()
+		}
+
+		if r.TimeUpdated != nil {
+			application["time_updated"] = r.TimeUpdated.String()
+		}
+
+		resources = append(resources, application)
+	}
+
+	if f, fOk := s.D.GetOkExists("filter"); fOk {
+		resources = ApplyFilters(f.(*schema.Set), resources, DataflowApplicationsDataSource().Schema["applications"].Elem.(*schema.Resource).Schema)
+	}
+
+	if err := s.D.Set("applications", resources); err != nil {
+		return err
+	}
+
+	return nil
+}