+++ oci/oci_dependency_graph.go
@@ -8,6 +8,7 @@ func initDependencyGraph() {
 	DependencyGraph = make(map[string][]string)
 
 	DependencyGraph["agreement"] = append(DependencyGraph["agreement"], "MarketplaceAcceptedAgreement")
+	DependencyGraph["application"] = append(DependencyGraph["application"], "DataflowInvokeRun")
 	DependencyGraph["application"] = append(DependencyGraph["application"], "FunctionsFunction")
 	DependencyGraph["asset"] = append(DependencyGraph["asset"], "CoreVolumeBackupPolicyAssignment")
 	DependencyGraph["autonomousContainerDatabase"] = append(DependencyGraph["autonomousContainerDatabase"], "DatabaseAutonomousDatabase")