+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Support `shape` property as customer input for VMDG.
 - Support Data Catalog Cloud Service.
+- Support for Dataflow Service.
 
 ## 3.59.0 (January 22, 2020)